Fix: Manual toggling of Lights not possible while Lights disabled
Previously the patch to disable traffic lights hooked into
RoadBaseAI.UpdateNodeFlags. But this method is not only called for
new junctions but when manual changing lights too.
As a result the patch reseted all changed back to no lights.
Now the patch hookes into RoadBaseAI.UpdateNode which is not called
for manual traffic light changes.